adding test on existing file instead of new file
