+++ api/v1beta2/awsmachine_types.go
@@ -226,12 +226,12 @@ type AWSMachineSpec struct {
 
 	// HostID specifies the Dedicated Host on which the instance should be launched.
 	// +optional
-	HostID *string `json:"hostId,omitempty"`
+	HostID *string `json:"hostID,omitempty"`
 
 	// Affinity specifies the dedicated host affinity setting for the instance.
 	// When affinity is set to Host, an instance launched onto a specific host always restarts on the same host if stopped.
 	// +optional
-	// +kubebuilder:validation:Enum:=Defailt;Host
+	// +kubebuilder:validation:Enum:=Default;Host
 	HostAffinity *string `json:"hostAffinity,omitempty"`
 }
